requirements.txt
pymupdf
streamlit_chat
langchain
streamlit
tiktoken
faiss-cpu
python-dotenv
chromadb
supabase